split indexer executor into writer and reader executor
